feat(util + net): common buffer util & buffer support for @lune/net
This commit implements buffer types are arguments and returns for
previously used string types for the @lune/net library.

Methods affected:
* net.request - Returns resp body as a buffer now
* net.serve - RequestConfig handlers can have a body of type string or
  buffer now
* net.socket - WebSocket stream data can now be sent as a string or
  buffer
